+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/io/DefaultFileIOFactory.java
@@ -23,6 +23,7 @@ import io.smallrye.common.annotation.Identifier;
 import jakarta.annotation.Nonnull;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import java.time.Clock;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -34,6 +35,8 @@ import org.apache.polaris.core.context.CallContext;
 import org.apache.polaris.core.persistence.PolarisResolvedPathWrapper;
 import org.apache.polaris.core.storage.AccessConfig;
 import org.apache.polaris.core.storage.PolarisStorageActions;
+import org.apache.polaris.core.storage.StorageAccessProperty;
+import org.apache.polaris.service.storage.StorageConfiguration;
 
 /**
  * A default FileIO factory implementation for creating Iceberg {@link FileIO} instances with
@@ -48,10 +51,17 @@ import org.apache.polaris.core.storage.PolarisStorageActions;
 public class DefaultFileIOFactory implements FileIOFactory {
 
   private final AccessConfigProvider accessConfigProvider;
+  private final StorageConfiguration storageConfiguration;
+  private final Clock clock;
 
   @Inject
-  public DefaultFileIOFactory(AccessConfigProvider accessConfigProvider) {
+  public DefaultFileIOFactory(
+      AccessConfigProvider accessConfigProvider,
+      StorageConfiguration storageConfiguration,
+      Clock clock) {
     this.accessConfigProvider = accessConfigProvider;
+    this.storageConfiguration = storageConfiguration;
+    this.clock = clock;
   }
 
   @Override
@@ -64,10 +74,14 @@ public class DefaultFileIOFactory implements FileIOFactory {
       @Nonnull Set<PolarisStorageActions> storageActions,
       @Nonnull PolarisResolvedPathWrapper resolvedEntityPath) {
 
+    // TODO
+    //    Boolean skipCredIndirection =
+    //        realmConfig.getConfig(FeatureConfiguration.SKIP_CREDENTIAL_SUBSCOPING_INDIRECTION);
+
     // Get subcoped creds
     properties = new HashMap<>(properties);
     AccessConfig accessConfig =
-        accessConfigProvider.getAccessConfig(
+        accessConfigProvider.getAccessConfigForCredentialsVending(
             callContext,
             identifier,
             tableLocations,
@@ -82,6 +96,34 @@ public class DefaultFileIOFactory implements FileIOFactory {
     properties.putAll(accessConfig.credentials());
     properties.putAll(accessConfig.extraProperties());
     properties.putAll(accessConfig.internalProperties());
+
+    // If no subscoped creds were produced, use system-wide AWS or GCP credentials if available.
+    if (accessConfig.credentials().isEmpty()) {
+      if (storageConfiguration.awsAccessKey().isPresent()
+          && storageConfiguration.awsSecretKey().isPresent()) {
+        // If no subscoped creds, use system-wide AWS credentials if available
+        properties.put(
+            StorageAccessProperty.AWS_KEY_ID.getPropertyName(),
+            storageConfiguration.awsAccessKey().get());
+        properties.put(
+            StorageAccessProperty.AWS_SECRET_KEY.getPropertyName(),
+            storageConfiguration.awsSecretKey().get());
+      } else if (storageConfiguration.gcpAccessToken().isPresent()) {
+        properties.put(
+            StorageAccessProperty.GCS_ACCESS_TOKEN.getPropertyName(),
+            storageConfiguration.gcpAccessToken().get());
+        properties.put(
+            StorageAccessProperty.GCS_ACCESS_TOKEN_EXPIRES_AT.getPropertyName(),
+            String.valueOf(
+                clock
+                    .instant()
+                    .plus(
+                        storageConfiguration
+                            .gcpAccessTokenLifespan()
+                            .orElse(StorageConfiguration.DEFAULT_TOKEN_LIFESPAN))
+                    .toEpochMilli()));
+      }
+    }
 
     return loadFileIOInternal(ioImplClassName, properties);
   }